Rename error message files and  function
